requirements_minimal.txt
# Minimal working requirements - Install in ORDER
# This avoids all conflicts by using compatible versions

# Step 1: Core dependencies (install first)
numpy==1.26.4
Pillow==10.4.0

# Step 2: Gradio and dependencies
gradio==4.44.0
gradio-client==1.3.0

# Step 3: Matplotlib
matplotlib==3.7.2

# Step 4: PDF processing (compatible versions)
pypdf==3.17.4  # Older version compatible with mineru
pdfminer.six==20221105
pypdfium2==4.30.0

# Step 5: Document processing
python-docx==1.1.2
markdown==3.6
openpyxl==3.1.5

# Step 6: LightRAG
lightrag-hku==1.4.5

# Step 7: LLM providers
openai==1.50.0
google-generativeai==0.8.5

# Step 8: Azure
azure-identity==1.17.1
azure-core==1.30.2

# Step 9: Utilities
python-dotenv==1.0.1  # Older stable version
aiohttp==3.9.5
tenacity==8.2.3
tiktoken==0.5.2
networkx==3.2.1
pydantic==2.5.3
nest_asyncio==1.6.0
pandas==2.2.0
tqdm==4.66.2

# Step 10: Additional
protobuf==4.25.3
grpcio==1.60.1
huggingface_hub==0.20.3
requests==2.31.0
typing-extensions==4.9.0